+++ packages/modal/src/index.spec.js
@@ -10,20 +10,7 @@ import * as rootIndex from './';
 
 describe('Index', () => {
   it('exports all components and utilities', async () => {
-    const exports = await getExports({
-      cwd: __dirname,
-      fileMapper: files => {
-        return files
-          .map(entry =>
-            entry
-              .replace(/\.js$/u, '')
-              .split('/')
-              .pop()
-          )
-          .filter(file => !/usePopper/u.test(file))
-          .sort();
-      }
-    });
+    const exports = await getExports({ cwd: __dirname });
 
     expect(Object.keys(rootIndex).sort()).toEqual(exports);
   });