+++ packages/selection/src/index.spec.js
@@ -7,6 +7,7 @@
 
 import { getExports } from '@zendeskgarden/react-testing';
 import * as rootIndex from './';
+import { composeEventHandlers, KEY_CODES } from '@zendeskgarden/container-utilities';
 
 describe('Index', () => {
   it('exports all components and utilities', async () => {
@@ -20,7 +21,8 @@ describe('Index', () => {
               .split('/')
               .pop()
           )
-          .filter(file => !/(ACTIONS|DIRECTIONS|getControlledValue)/u.test(file))
+          .filter(file => !/ACTIONS|DIRECTIONS/u.test(file))
+          .concat(Object.keys({ composeEventHandlers, KEY_CODES }))
           .sort();
       }
     });