+++ packages/breadcrumb/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/zendeskgarden/react-containers/compare/@zendeskgarden/container-breadcrumb@0.1.3...@zendeskgarden/container-breadcrumb@0.2.0) (2019-05-06)
+
+
+### Features
+
+* **utilities:** add new `@zendesgarden/container-utilities` package ([#30](https://github.com/zendeskgarden/react-containers/issues/30)) ([fa901cc](https://github.com/zendeskgarden/react-containers/commit/fa901cc))
+
+
+
+
+
 ## [0.1.3](https://github.com/zendeskgarden/react-containers/compare/@zendeskgarden/container-breadcrumb@0.1.2...@zendeskgarden/container-breadcrumb@0.1.3) (2019-03-13)
 
 **Note:** Version bump only for package @zendeskgarden/container-breadcrumb